refactor(board): move board files into components/game folder
