+++ frontend/src/components/game/boardHelpers.tsx
